requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pinecone
anthropic
sentence-transformers
tqdm
